hwc: change the behavior of composerLayer [2/3]
PD#SWPL-18099

Problem:
When playing a 4k video and press back in YouTube, SurfaceView will not
be destroyed. And in composer service process, there is a composerLayer.
CompserLayer has a ComposerHandleCache which will hold the graphic
buffer handles cached in it and lead to ion orphaned memory.

Solution:
Change the behavior of composerLayer.

Verify:
Verify on newton

Change-Id: I42f952c61826588c283bd3363fbfef6df444f7f6
Signed-off-by: baocheng sun <baocheng.sun@amlogic.com>